clean up; round 2 hex output, lock cv inspection, sigpipe resiliance

+++ util.h
@@ -5,7 +5,7 @@
 extern int gverbose;
 ssize_t bufferfill(int fd, u_char *__restrict dest, size_t size, int charmode);
 void stopwatch_start(struct timespec *t);
-int stopwatch_stop(struct timespec *t, int whisper_channel);
+u_long stopwatch_stop(struct timespec *t, int whisper_channel);
 #ifdef CHAOS
 int chaos_fail();
 #endif
